move tests to cassandra
