+++ atlasdb-cassandra-integration-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraGetCandidateCellsForSweepingTest.java
@@ -16,13 +16,19 @@
 
 package com.palantir.atlasdb.keyvalue.cassandra;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import org.junit.ClassRule;
+import org.junit.Ignore;
+import org.junit.Test;
 import org.mockito.Mockito;
 import org.slf4j.Logger;
 
 import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceConfigManager;
 import com.palantir.atlasdb.containers.CassandraContainer;
 import com.palantir.atlasdb.containers.Containers;
+import com.palantir.atlasdb.encoding.PtBytes;
+import com.palantir.atlasdb.keyvalue.api.ImmutableCandidateCellForSweeping;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.impl.AbstractGetCandidateCellsForSweepingTest;
 
@@ -37,5 +43,49 @@ public class CassandraGetCandidateCellsForSweepingTest extends AbstractGetCandid
                 CassandraKeyValueServiceConfigManager.createSimpleManager(CassandraContainer.KVS_CONFIG),
                 CassandraContainer.LEADER_CONFIG,
                 Mockito.mock(Logger.class));
+    }
+
+    @Test
+    public void returnCandidateIfPossiblyUncommittedTimestamp() {
+        new TestDataBuilder().put(1, 1, 10L).store();
+        assertThat(getAllCandidates(conservativeRequest(PtBytes.EMPTY_BYTE_ARRAY, 40L, 5)))
+                .containsExactly(ImmutableCandidateCellForSweeping.builder()
+                        .cell(cell(1, 1))
+                        .sortedTimestamps(new long[] { 10L })
+                        .isLatestValueEmpty(false)
+                        .numCellsTsPairsExamined(1)
+                        .build());
+    }
+
+    @Test
+    public void returnCandidateIfTwoCommittedTimestamps() {
+        new TestDataBuilder().put(1, 1, 10L).put(1, 1, 20L).store();
+        assertThat(getAllCandidates(conservativeRequest(PtBytes.EMPTY_BYTE_ARRAY, 40L, 30)))
+                .containsExactly(ImmutableCandidateCellForSweeping.builder()
+                        .cell(cell(1, 1))
+                        .sortedTimestamps(new long[] { 10L, 20L })
+                        .isLatestValueEmpty(false)
+                        .numCellsTsPairsExamined(2)
+                        .build());
+    }
+
+    // TODO(nziebart): check that this is ok to fail, and delete
+    @Ignore
+    @Test
+    public void doNotReturnCandidateWithCommitedEmptyValueIfConservative() {
+        new TestDataBuilder().putEmpty(1, 1, 10L).store();
+        assertThat(getAllCandidates(conservativeRequest(PtBytes.EMPTY_BYTE_ARRAY, 40L, 30))).isEmpty();
+    }
+
+    @Test
+    public void returnCandidateWithCommitedEmptyValueIfThorough() {
+        new TestDataBuilder().putEmpty(1, 1, 10L).store();
+        assertThat(getAllCandidates(thoroughRequest(PtBytes.EMPTY_BYTE_ARRAY, 40L, 30)))
+                .containsExactly(ImmutableCandidateCellForSweeping.builder()
+                        .cell(cell(1, 1))
+                        .sortedTimestamps(new long[] { 10L })
+                        .isLatestValueEmpty(true)
+                        .numCellsTsPairsExamined(1)
+                        .build());
     }
 }
